Handlebars helpers, IronRouter, and new Router instance auto created.
This is so that a router instance is automatically created when the package
is included. It's assigned to the Router variable. This way, users can get
started right away without worrying about creating an instance. All Handlebars
helpers use the global Router instance. The internal router constructor was
changed to IronRouter.