Allow Router.go to take a route name or a path.
The go method can now take a route name or a path. If the user provides a route
name, we treat the second parameter as params and the third as state. If the
user passes a path, we treat the second parameter as state.

We decide whether the first parameter is a path by checking whether it starts
with a forward slash or http. If it doesn't we assume it's a route name.