Location shouldn't invalidate deps unless state has actually changed.
This bug was discovered in the latest Canary build of Chrome. They changed the
behavior of initial page load with respect to the popstate event. Previously,
the popstate event was not fired when the page initially loads. Now it does
fire. So if a client called Location.state(), the function would get fired twice
on initial page load.

Now the deps only invalidate if the state actually changed.