Initial stab at delayed inheritance. Needs feedback.
The flushing mechanism works as follows:

RouteController.create() delays inheritance until after all initial code has
loaded and executed by putting the inheritance code in a callback to
RouteController.onCreateFlush(...).

The Router in turn, doesn't start until the createFlushQueue is drained, which
happens automatically at the right time if anything is added to it using the
onCreateFlush. The router does this by putting its start code in a callback to
RouteController.afterCreateFlush(...).

This works similarly to the Deps.afterFlush and Deps.onFlush methods. One
limitation of this approach is that there is a slight delay before the router
renders. It might be more performant to control load order by putting code in
packages. But this seems a lot to ask of app authors. But it might be years
before Meteor releases a load ordering mechanism.