Fix route computation bug and data isolation issue in render.
1. Data isolation issue:

Inside the render method of controller is where we call the data function if
it's a function. Previously, the function call was not isolated, so if the deps
invalidated the entire route would be rerun instead of just the data function.
That kind of defeats the point of having a data function. So I added an autorun
around the call to wait (which encloses the onReady and onWait methods. This has
the added benefit of making wait work even if reactive: false option is set on
the route.

2. Route computation bug

The bug could be triggered by the following example click event handler:

'click': function (e, tmpl) {
 Posts.update(...);
 Router.go('anotherRoute');
}

Problem:
The problem is that the update invalidates the current route's computation, but
before that computation is flushed, the new route is run. The onRun method in
ClientRouter checks for a previous computation and finds one. Then it said if
the previous computation is marked as invalid, run the new route in an
invalidation callback. But since the computation is currently marked as invalid
(but not stopped) the onInvalidate function is run immediately (before the stop
method is even called). This causes the new route to run right away. But then we
call stop on the previous computation, which has now been reset to the new
route. And then the previous route handler re-runs, rendering the template old
template to the page.

Fix:
If we find a previous computation stop it immediately. Then in an afterFlush
callback run the new route. This will cause the new computation not to be a
child of the old one and therefore not be stopped.