Fix standard hooks...
You no longer get these hooks for free. You need to ask for them. You can pass a
string name for the hook. For example:

// run the loading hook for all routes as a before hook
Router.before('loading');

// run the notFound hook for all routes as a before hook
Router.before('notFound');

Remember that the order of hooks matters. In this example, you want the notFound
hook to come last since we want to show a loading indicator until the controller
is ready. The notFound hook checks whether the data value is falsy and if so,
renders the notFound template. But we don't want to do that until the data is
actually ready!