Fix up the data API on RouteController and fixes...
This makes the `this.data()` and this.data({title: 'example'})` api more
consistent and predictable. When you define a controller or route you can
provide a data property like this:

data: function () {
 return Items.findOne({_id: this.params._id})
}

Now, you can call this.data({...}) to set the data value at a later time, or
this.data() to get the value from inside one of your controller functions.