Finish iron-layout integration.
@tmeasday,

Want to do a sanity check on this and make sure IR still works for you? The
expectation is that it should work as it did before (maybe with a few bugs
fixed) but uses the new iron-layout under the hood. I think if this works okay
we can do a patch release.

Next up I'll work on the larger refactor. I'll try to do this in stages so that
we can have a clear discussion around each thing and separate R&D type
improvemenets from general cleanup and API hardening.