fix layout rendering bug with issue #724
There were two bugs which were fixed primarily by using the updated
iron-layout and iron-dynamic-template packages. I backported the
packaging release work and will do new releases of these packages.

There were actually two bugs:

1) template setter in iron-dynamic-template could create a dependency if
the parameter was undefined and the last value was also undefined. this
was fixed in the packaging branch but not backported.

2) the second bug could be reproduced by doing the following in your
action function:

this.layout(undefined);
this.layout('MyLayout');
this.render('sometemplate', {to: 'someregion'});

The rendering would happen on the wrong layout region. This has been
fixed in the newer iron-layout.