Allow meteor apps in subdirectories: /app/route1
If the project is configured to run in a sub-path of a domain, eg. http://www.xyz.com/app1/ via the meteor environment variable ROOT_PATH, we can remove that prefix from the path and then check if it matches a route.

That should allow for multiple instances of an app on a single domain.