reverting default template name converter to class case...
fixes #953

@tmeasday, sorry about this. But this is how all the examples in the
guide look, how it's taught on EM and also how the em command line tool
works by default. So I think quite a few people will be using the class
case and this change will cause confusion. I will add a note to the
README explaining how to change the template name converter to the
identity. I'll do this on the next commit.