remove EJSON.clone of data context in urlFor, pathFor and linkTo
fixes #965

Note, we still clone individual attributes. But for the main data
context we'll just create a new object and merge into it. That we we're
assigning new properties to the new object vs. mutating the original
data context. But we'll avoid the circular reference loops that cause
problems for EJSON.clone.