update tom's isHandled logic to use thisArg instead of params...
Fixes links to static assets, and allows for not found routes to be
rendered on the client. Does not handle the correct http status code
setting on server.

See:
https://github.com/EventedMind/iron-router/commit/9c642dc6d6501dbee4c1f068efeaa6ad74873fb9
https://github.com/EventedMind/iron-router/issues/1051

@tmeasday,

Your initial state solution looks good. I made a few modifications so
that we don't need to pass parameters around in the "done()" or "next()"
functions. These functions have a contract signature that comes from
iron:middleware-stack. See line 235 in lib/middleware_stack.js. So
overloading the parameter signature with a second parameter of "handled"
seems like a recipe for confusion. Instead, I just use the controller's
isHandled() function.

Next up is seeing if we can set the correct HTTP status code on the
server for routes that are unknown. I think this should be a PR to
Meteor.